fix: 상세페이지 텍스트 렌더링 오류 수정정
